Add CommandLine class to parse command line

+++ framework/run_manager.rb
@@ -88,19 +88,25 @@ module PACKMAN
           cmd_str << build_helper.wrap_flags(language, CompilerHelper.default_flags(language, compiler))
         end
       end
-      cmd_str << " 1> #{ConfigManager.package_root}/stdout 2> #{ConfigManager.package_root}/stderr"
+      if not PACKMAN::CommandOptions.has_option? '-d'
+        cmd_str << " 1> #{ConfigManager.package_root}/stdout 2> #{ConfigManager.package_root}/stderr"
+      end
       print "#{PACKMAN::Tty.blue}==>#{PACKMAN::Tty.reset} #{PACKMAN::Tty.truncate("#{cmd} #{args.join(' ')}")}\n"
       system cmd_str
       if not $?.success?
-        PACKMAN.report_error "Failed to run the following command:\n"+
-          "PATH: #{FileUtils.pwd}\n"+
-          "Command: #{cmd_str}\n"+
-          "Return: #{$?}\n"+
-          "Standard output: #{ConfigManager.package_root}/stdout\n"+
-          "Standard error: #{ConfigManager.package_root}/stderr\n"
+        info =  "PATH: #{FileUtils.pwd}\n"
+        info << "Command: #{cmd_str}\n"
+        info << "Return: #{$?}\n"
+        if not PACKMAN::CommandOptions.has_option? '-d'
+          info << "Standard output: #{ConfigManager.package_root}/stdout\n"
+          info << "Standard error: #{ConfigManager.package_root}/stderr\n"
+        end
+        PACKMAN.report_error "Failed to run the following command:\n"+info
       end
-      FileUtils.rm("#{ConfigManager.package_root}/stdout")
-      FileUtils.rm("#{ConfigManager.package_root}/stderr")
+      if not PACKMAN::CommandOptions.has_option? '-d'
+        FileUtils.rm("#{ConfigManager.package_root}/stdout")
+        FileUtils.rm("#{ConfigManager.package_root}/stderr")
+      end
     end
   end
